[FIX] Quick-win audit fixes: typos, bare except, sphinx underlines
- bdd_oia/main.py: --segment_size type=int -> type=float (default 0.01)
- simple_bridge.py: replace bare except with (TypeError, ValueError)
- basic_nn.py: fix 'accuray' typo in log message and 'save_parma_dic' var
- docs/API/*.rst: fix title underlines (5 files); list members explicitly
  for ablkit.reasoning and ablkit.utils so newly-added symbols surface
- docs/Intro/{Basics,Quick-Start,Reasoning}.rst: fix title underlines
- docs/Examples/HWF.rst: '14 classes' -> '13 classes' (1-9 + 4 ops);
  fix 'illstrations' typo; fix 'alongwith' -> 'along with'
- docs/Examples/MNISTAdd.rst: fix 'alongwith' -> 'along with'

+++ examples/bdd_oia/main.py
@@ -48,7 +48,7 @@ def get_args():
         "--loops", type=int, default=2, help="number of loop iterations (default : 2)"
     )
     parser.add_argument(
-        "--segment_size", type=int, default=0.01, help="segment size (default : 0.01)"
+        "--segment_size", type=float, default=0.01, help="segment size (default : 0.01)"
     )
     parser.add_argument("--save_interval", type=int, default=1, help="save interval (default : 1)")
     parser.add_argument(
